requirements.txt
aiohttp
watchdog
asyncio
python-dotenv
requests
discord
lxml

